+++ packages/shared/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @step-func-emailer/shared
+# @mailshot/shared
 
 ## 0.2.0
 
 ### Minor Changes
 
-- [`72d23af`](https://github.com/mdwt/step-func-emailer/commit/72d23af8e07ab379c794d2918990f7a517a5fc85) Thanks [@mdwt](https://github.com/mdwt)! - Publish framework packages to npm and add create-step-func-emailer CLI
+- [`72d23af`](https://github.com/mdwt/mailshot/commit/72d23af8e07ab379c794d2918990f7a517a5fc85) Thanks [@mdwt](https://github.com/mdwt)! - Publish framework packages to npm and add create-mailshot CLI